requirements.txt
moviepy>=1.0.3
openai-whisper>=20230314
numpy>=1.20.0
torch>=1.10.0